Small improvements and fixed #10

+++ examples/globals/eq.cpp
@@ -65,7 +65,7 @@ int rule_operator_eq_callback(struct rules_t *obj, int a, int b, int *ret) {
       case VFLOAT: {
         struct vm_vfloat_t *na = (struct vm_vfloat_t *)&obj->varstack.buffer[a];
         struct vm_vfloat_t *nb = (struct vm_vfloat_t *)&obj->varstack.buffer[b];
-        if(fabs((float)na->value-(float)nb->value) < EPSILON) {
+        if(fabsf((float)na->value-(float)nb->value) < EPSILON) {
           out->value = 1;
         } else {
           out->value = 0;
